backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
python-jose
passlib[bcrypt]
python-dotenv
email-validator
